This is a simple Python compiler that you can use to compile your python programs to .exe files.
To use it, make sure that this file and the one which you want to compile are present in the same folder. 
Also, install 'pyinstaller' if you haven't on your computer.
To do so, simply enter the command inside the brackets ( pip install pyinstaller ) on your computer's terminal. 
After ensuring everything mentioned above, open your terminal and enter the location of the folder. Make sure you follow the syntax of your respective operating system to do so.
Then, enter this command: python compile.py "Your_file_name.py".
After this, wait for the terminal to compile your program (the time period depends on the size of your code...).
Done! Now your .exe file will be created inside a dist folder in the same folder where your program and this compiler are located.
Thank you! Enjoy! 🎉